Playing around, hoping to optimise things a little.
git-svn-id: svn+ssh://circa.ecs.vuw.ac.nz/u/staff/djp/projects/repository/tuttepoly/trunk@850 cab769db-1af3-0310-8bf3-f2d376eea66d

+++ src/poly/simple_poly.hpp
@@ -9,10 +9,10 @@
 
 class term {
 public:
-  unsigned int xpower;
-  unsigned int ypower;
+  unsigned short xpower;
+  unsigned short ypower;
 
-  term(unsigned int x, unsigned int y) : xpower(x), ypower(y) {}
+  term(unsigned short x, unsigned short y) : xpower(x), ypower(y) {}
 
   bool operator<(term const &t) const {
     return (xpower < t.xpower) || (xpower == t.xpower && ypower < t.ypower);
